Refactor for last tool result
