+++ src/github/quickPicks.ts
@@ -385,7 +385,7 @@ export async function getLabelOptions(
 	const labelPicks = newLabels.map(label => {
 		return {
 			label: label.name,
-			description: label.description,
+			description: label.description ? label.description : undefined,
 			picked: labels.some(existingLabel => existingLabel.name === label.name),
 			iconPath: DataUri.asImageDataURI(Buffer.from(`<svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
 				<rect x="2" y="2" width="12" height="12" rx="6" fill="#${label.color}"/>
