+++ src/lm/tools/tools.ts
@@ -7,14 +7,15 @@
 import * as vscode from 'vscode';
 import { RepositoriesManager } from '../../github/repositoriesManager';
 import { DisplayIssuesTool } from '../displayIssuesTool';
+import { ChatParticipantState } from '../participants';
 import { ConvertToSearchSyntaxTool, SearchTool } from '../searchTools';
 import { IssueTool } from './issueTool';
 import { SuggestFixTool } from './suggestFixTool';
 
-export function registerTools(context: vscode.ExtensionContext, repositoriesManager: RepositoriesManager) {
+export function registerTools(context: vscode.ExtensionContext, repositoriesManager: RepositoriesManager, chatParticipantState: ChatParticipantState) {
 	registerIssueTool(context, repositoriesManager);
 	registerSuggestFixTool(context, repositoriesManager);
-	registerSearchTools(context, repositoriesManager);
+	registerSearchTools(context, repositoriesManager, chatParticipantState);
 }
 
 function registerIssueTool(context: vscode.ExtensionContext, repositoriesManager: RepositoriesManager) {
@@ -25,8 +26,8 @@ function registerSuggestFixTool(context: vscode.ExtensionContext, repositoriesMa
 	context.subscriptions.push(vscode.lm.registerTool('github-pull-request_suggest-fix', new SuggestFixTool(repositoriesManager)));
 }
 
-function registerSearchTools(context: vscode.ExtensionContext, repositoriesManager: RepositoriesManager) {
+function registerSearchTools(context: vscode.ExtensionContext, repositoriesManager: RepositoriesManager, chatParticipantState: ChatParticipantState) {
 	context.subscriptions.push(vscode.lm.registerTool('github-pull-request_searchSyntax', new ConvertToSearchSyntaxTool(repositoriesManager)));
 	context.subscriptions.push(vscode.lm.registerTool('github-pull-request_doSearch', new SearchTool(repositoriesManager)));
-	context.subscriptions.push(vscode.lm.registerTool('github-pull-request_renderIssues', new DisplayIssuesTool(repositoriesManager)));
+	context.subscriptions.push(vscode.lm.registerTool('github-pull-request_renderIssues', new DisplayIssuesTool(chatParticipantState)));
 }
